pyproject.toml
[project]
name = "ssu-isat"
version = "0.1.0"
description = "Educational project for SSU ISaT course"
authors = [
    { name = "Artemiy Granat", email = "artemiigranat@gmail.com" }
]
dependencies = [
    "fastapi>=0.109.0",
    "pydantic>=2.5.3",
    "uvicorn>=0.25.0",
    "bs4>=0.0.2",
    "httpx>=0.26.0",
    "asgi_correlation_id>=4.3.0",
    "databases>=0.8.0",
    "tenacity>=8.2.3",
    "imagehash>=4.3.1",
    "pillow>=10.2.0",
    "apscheduler>=3.10.4",
    "rembg>=2.0.53",
    "python-multipart>=0.0.9",
    "scipy>=1.12.0",
    "torchvision>=0.17.1",
    "openai-clip>=1.0.1",
    "setuptools>=69.1.1",
    "asyncpg>=0.29.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.1.13",
    "pre-commit>=3.5.0",
]

[tool.hatch.metadata]
allow-direct-references = true

# TODO: format file

[tool.hatch.build.targets.wheel]
packages=["backend/src", "scraper/src", "image-processer/src", "image-blender/src", "color-search/src", "shared"]

[tool.ruff]
line-length = 79
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["I", "E4", "E7", "E9", "F"]
ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
